Crate sey_currency_converter_pwa

Modules

currdb_config_mod

currdb_currency_mod

currdb_currency_used_mod

currdb_manual_rates_mod

currdb_mod

fetch_rates_mod

idbr_imports_mod

idbr_mod

page_input_currency_mod

page_input_currency_used_mod

page_main_mod

page_manual_rates_mod

page_modal_about_mod

page_output_currency_mod

page_output_currency_used_mod

utils_mod

utils_mod.rs

web_sys_mod

helper functions for web_sys, window, document, dom, console, local_storage, session_storage,…

Macros

on_click

Simple macro to set listener of on_click events to an element_id.
fn with 1 arg(element_id): on_click!(element_id, function_ident)

on_keyup

Simple macro to set listener of onkeyup events to an element_id. on_keyup!(“regex_text”, run_regex)

row_on_click

list contains rows, every row item needs a separate event handler the element Id is concatenation of element_prefix plus the row_number

Functions

__wasm_bindgen_generated_wasm_bindgen_start

To start the Wasm application, wasm_bindgen runs this functions

wasm_bindgen_start

To start the Wasm application, wasm_bindgen runs this functions